issue #14: wip
