+++ package.json
@@ -54,7 +54,7 @@
 		"object-observer": "^4.0.2"
 	},
 	"devDependencies": {
-		"eslint": "^7.12.0",
+		"eslint": "^7.12.1",
 		"just-test": "2.3.2",
 		"uglify-es": "^3.3.9",
 		"fs-extra": "^9.0.1"
